Move RealmObject utility extension functions to RealmObjectUtil
