+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/query/SingleQuery.kt
@@ -2,13 +2,13 @@ package io.realm.internal.query
 
 import io.realm.RealmObject
 import io.realm.asFlow
-import io.realm.hasSameObjectKey
 import io.realm.internal.InternalDeleteable
 import io.realm.internal.Mediator
 import io.realm.internal.Observable
 import io.realm.internal.RealmReference
 import io.realm.internal.RealmResultsImpl
 import io.realm.internal.Thawable
+import io.realm.internal.hasSameObjectKey
 import io.realm.internal.interop.ClassKey
 import io.realm.internal.interop.Link
 import io.realm.internal.interop.NativePointer
